[privacy sandbox] Use appropriate AndroidVersion
Current sdklib.AndroidVersion is used for compileSdk, targetSdk
and minSdk in PrivacySandboxSdkVariantScope.
However, sdklib.AndroidVersion is only needed for
compileSdk (containing major and minor versions). So targetSdk and
minorSdk can use variant.AndroidVersion. This is important as both major
and minor version version codes are included in the API string from API >=
36.

Fixes: 411666019
Test: Existing privacy sandbox tests (no behaviour changes)
Change-Id: I2103b35334e54128e80ebe72e493afd7f0e36e55