Fix manifest merge error when the API version is major.minor
The merger was attempting a parseInt on a floating point number (36.0),
which resulted in an error.

Because we were not able to merge the manifest, we were not able to
detect if a manifest had the "uses-feature watch" required for wear
apps.

This CL changes the logic to use AndroidVersion.fromString which should
handle such cases.

Fixes: 409968284
Test: added regression test
Change-Id: I0c68dc7152da782afa119857748f079416266224