Make @FlaggedApi check not depend on the flags class
Until now the lint check for @FlaggedApi required the Flags class
to also be on the classpath (because it attempts to resolve the linked
class and if not found, gives up). However, there are cases where
users of an API do not have the flags classes on the classpath, so
there would be no lint violations raised.

Bug: Not filed
Test: Unit test included
Change-Id: I7dda9845891f0b14c0110f3d6b0a2a33f0ff1626