Allow marking Context and Project as tests
Adds Project.isTestProject so an entire Project can be marked as a
test. Moves JavaContext.isTestSource up to Context.isTestSource so
other files (like resources and manifests) can be marked as test files.
Also fixes a bug (and related challenges) with marking a source file as
both a test and generated source file in a project.xml file.

Note that the default behavior of Project.isTestProject is to return
true if there are test source folders and no other source folders, and
null (unknown) otherwise. ManualProject overrides this to be explicit.

Context.isTestSource is initialized to true if its Project is a test.

Note that Lint still visits XML files that end up being marked as test
files without requiring the TEST_SOURCES scope or checkTestSources flag,
unlike with test Java/Kotlin sources. This means we have not changed the
current behavior, for now, but we could change this in the future.

Bug: 359915216
Test: Added to ProjectInitializerTest
Change-Id: I932d89363ba0be29aeebe270a487373d44dc7625