added test suites for AndroidProject model
The model is very simple at this stage, it only contains the test
task name to run the test suite and the list of configured junit
test engines to do the actual test execution.

Change the test task implementation to not initialize the device
provider when TESTED_APKS are not requested in the JUnit engine
DSL Configuration. We might revisit this later to maybe add an
explicit flag in the test suite declaration.

The sources information will be added to BasicVariant model in a
subsequent CL.

Test: Added integration test.
Bug: N/A
Change-Id: I9140627c4c6dc5969065d639835d168c39200630