Patch perfetto to support Bazel 8
Perfetto is added using http_archive so it can be patched
with https://github.com/google/perfetto/pull/1300.
This change is required to make it work on Bazel 8.

Perfetto is patched instead of being upgraded because
the latest version pulls in new dependencies that are not
easily built. For example, libexpat is a new dep that
requires a newer glibc version than whats in prebuilts.
Newer perfetto also causes test failures, so some code
may need updating.

Bug: 409124110
Change-Id: I1278a8e2d4652472d084a06849178b8d11e001df