Fix bug in RichVersion#toIdentifier
Somehow generating an identifier for a RichVersion with a strict
requirement but no specified preferred version was both untested and
buggy in the most obvious way.  Address both of these problems.

This also makes some minor changes to the behaviour of Lint's
GradleDetector.  The change in the textual message is broadly
neutral (possibly a bit confusing in this specific case, but
consistent with what will happen if there is a preferred version in
the version specifier); the change in the quick fix is reasonable,
because there should be no correlation between one version needing to
be strictly specified, and the subsequent version requiring a strict
specification.

Bug: none filed
Test: added, existing test adjusted
Change-Id: I50beb66f39849360f5ffed29ce5357a6b42d81f8