Fix DeviceTestBuilder.enable
The filtering of device tests was done incorrectly forcely
disabling tests that were not in the default build type.
Instead, we now just use the enable flag without any other
conditions and use a defined step after the DeviceTestBuilder
createion to initialize the enable flag correctly.

That way, if users change the flag value, it's done after the
initialization.

Test: added a regression test
Fixes: 390602110
Change-Id: I1f08b921b8aa792fd20812fcb2527ef7e50f578e