Provide a static `find` for WellKnownMavenArtifactId
This isn't desperately well-founded, in that it would be fairly easy
to forget to add a suitable initializer block (or equivalent) for an
implementation class.  However, it is mitigated by having a sealed
interface at the top of the heirarchy, so any mistakes are in fact
ours.  (This is probably a temporary measure while converting project
system APIs away from GradleCoordinate, so the lack of soundness is
acceptable).

Bug: 279886738
Test: added
Change-Id: If151a297460ac42c38927710e3b6b40d1d879cc3