Support multiple source types for test suites.
Add the ability to declare the types of sources used for the
test suite. So far, you can declare three types :
- assets for static declarative files
- hostJar for kotlin sources that will run on host machine
- testApk for android source set to produce test APKs.

So far, only host is supported and you can only have one instance
of each type. The model has been also updated to move the sources
declaration from the Variant objects back to the BasicAndroidProject
model. The variant object just have a reference of the test suites
that applies.

Test: Adapted existing tests.
Bug: N/A
Change-Id: I4ad9f33a6bfcfc955b4288f8bd80515889391075