Introduce BooleanOption for new DSL implementation
The legacy implementations delegate to the new ones, and so just switch
to exposing the new interface as the 'android' extension.

Some more internal cleanup to follow - many internals of AGP depend on
the legacy DSL implementations. To avoid blocking this on all that
cleanup, when setting
    android.newDsl=true
AGP will still construct an internal instance of the legacy
implementation which is not exposed as an extension, but is used
internally.

Bug: 418804641
Test: Added smoke test, more to follow
Change-Id: I8c66dc1999f8be54a6ad11d8468f4acbeca30582