Update build.gradle generation in Auto template
This template is special in that it generates a build.gradle file within
the template-impl package rather than the android-npw package, where
most module creation is done. As a result, it doesn't have the updates
made previously to support minor versions (and new Gradle syntax, and
KTS support.)

Copy from sharedMacros.kt to fix the compile/min/target SDK properties.

Bug: 419193356
Bug: 315945396
Test: generated and built template
Change-Id: Iac741fcef1ace3f48dd86c653cda4739c80cf8fe