Tweak test target of //tools/base/repostory:tools.tests in coverage
This test target was converted from a `coverage_java_test` to a
`kotlin_test` in ag/33315624. This causes the underlying test target
name to change from `...tools.tests` to `...tools.tests.test`.

The coverage graph validation check in the linux builder won't detect
this sort of thing because `//tools/base/repository:tools:tests` is
still a sufficient target for *running* the test. Just not for finding
it in the test results.

Bug: N/A
Test: N/A
Change-Id: I4822372babfdd286aedafbec19490fcd9b711019