Use max flaky retries in coverage builds
Since we're only running coverage nightly now flaky tests are a lot more
annoying and wasteful. Turning on flaky retries will make it far less
likely for a night's run to get wasted by a flake. It will also make it
much easier to identify if the build is actually broken because
transient flakes will not be polluting the build history.
Using max retries (10) shold be ok since we only do one run a day.

Bug: N/A
Test: N/A
Change-Id: If0936bbe8fca8a5c63adb12574adb1a43f791efd