Revert including API level in previews
This reverts the effects of Ic5de6c46da98a917 and I29e00ec291070eef, but
keeps and updates most of the tests they added.

Instead of having API levels prefixed to codenames, we'll have a minor
version suffix on the codenames in package metadata (though not in
system properties).

Bug: 402544627
Test: Updated
Change-Id: I41d88855adb1ee333cff9493066e7dfe14fdacc7