Replace `ThreadDetector` and `WrongThreadInterproceduralDetector`
Within the IDE, the new thread detector doesn't have access beyond the
active file, so should not be too slow. The detector also has been
tested on several larger modules in commandline, finishing in
reasonable time. As far as tests go, the new detector subsumes
previous ones, and covers many more features not previously considered.

Therefore we replace the old thread detector with the new one in
Canaries for feedback.

Bug: 419864689
Bug: 419866848
Test: existing
Change-Id: I459c5163352d681244a98f879c5a127ffaa777e3