Handle unsigned literals
The function `ofLiteral` returns the type of constant literals, and is
supposed to be exhaustive on the fixed set of literals defined in the
language, but wasn't handling unsigned integers.

This change also suppresses errors in the analysis in production, per
unit (e.g. method summary, or summary instantiation), to avoid
bringing down all of Lint.

Bug: n/a
Test: new
Change-Id: Icf985a0c7f8f6c30c87900d327e46a014cc35044