Generate proxy interfaces for kotlin lambdas
VM-generated proxy instances can expose a public interface based on a
set of interface classes, but cannot inherit from a superclass or be
cast to any particular type.

The coroutine library performs runtime type checks for certain logic,
requiring Live Edit proxy instances to be typed 'properly' in order for
proper program behavior to occur.

To solve this problem, Live Edit generates a set of proxy interfaces for
all possible lambda superclass + functional interface combinations when
studio is built. At runtime, when a lambda proxy interface needs to be
created, an instance of the matching custom proxy interface is used
instead of an instance of java.lang.reflect.Proxy.

Bug: 269627284
Test: LambdaSuperTest
Change-Id: I97ee220da53722bf677e9e0ddfabb0c2ae10e3e9