intellij_plugin: limit to supported platforms
The intellij_plugin() macro declares a studio_plugin() target,
which gets built as a standalone target in the default configuration
when someone runs `bazel build //...`. This means that 'aiplugin' gets
configured against the 'studio-sdk' platform (for example) even
though it doesn't need to be---and it leads to some minor issues, e.g.
Bazel expects all select-expressions to have a clause for studio-sdk.

Bug: unblocks b/424003399
Test: Bazel build
Change-Id: I9a7996867d83762cc5a3d9b0d41bff5ad7415570