Widen symbolic arguments on instantiation
The implementation previously only summarized results and not calls,
so could diverge on recursion that kept growing symbolic arguments.
This change widens symbolic values on instantiation to ensure they
don't grow unbounded.

In addition to `testRecursion_accum_virtual` reproducing the problem,
this change also tests polymorphic recursion, and a few tests on an
interpreter for a small higher-order language, which typically make
good mini stress tests for static analyses, ensuring convergence in
reasonable time with the expected precision.

Bug: n/a
Test: new
Change-Id: I28addceee2b696827bac34c200f92f4b013cc9e4