Update IntelliJ versions used in bazel
For 2025.1, there were breaking changes in how Jewel is bundled in the
IDE from 2025.1 to 2025.1.1. Rather than introduce another sdkcompat
layer for minor updates (and additional complexity of bundling different
amounts of Jewel based on the minor version), we update to 2025.1.1.

For 2025.2, there were several breaking changes in the EAPs since the
initial 2025.2 EAP was added. Here, it allows for less diffs by updating
this one to a newer EAP.

Bug: 378750746
Change-Id: I21d65445722b817bec5bea72ba5530bfbec422a8