adblib: Initial value of isWaitingForDebugger
When using the JDWP process properties collector,
keep `isWaitingForDebugger` value "empty" until we
known 1) either the value is `true` (through a DDM WAIT packet)
or 2) we are done collecting properties (in which case the value
is `false`).

This change allows consumers of JdwpProperties to know
whether `isWaitingForDebugger` is not yet known,
or `false` for sure.

Test: Updated existing tests
Bug: n/a
Change-Id: I4d3b4ed0b98274cae1f8182889fe2f0e48c87349