Fix missing resources added via a task provider in KMP
Resources added via the variant api (addGeneratedSourceDirectory) are
not properly processed and missing from the AAR. The reason was that we
would look for a ResourceSet named "generated" when computing resource
sets for merging. (DependencyResourcesComputer.compte() is the method in
question). We now create the directory entry upon creation of the res
sources so that it exists when the resource directories are added.

Bug: 425594733
Test: KotlinMultiplatformGeneratedResourcesTest
Relnote: Fix missing android resources added via Variant API using a task provider in AGP-KMP plugin
Change-Id: I3ed4967715851dc9af0f8e3d5226ae41c0f93e04