Determine isKMP by dependsOn or non-JVM platform
Also fix config cycle bug in LintDriver that can occur if two modules
have the same root directory.

Also, make it so one of the SCANNING_FILE events is now fired inside a
read action, as this is how it is normally done, and this allow tests to
observe withMultiplatformLightClassSupport in a listener. We don't
actually use this in the added tests, but it was discovered while
prototyping different ways to test that KMP support is working.

Bug: 409337070
Bug: 427821470
Test: Added
Change-Id: I6cce967939533b2080a7713536ffe0a9c621cb83