[fused lib][res] Rewrite only refs for resources defined once fused
When rewriting packages of R class resource references of classes that are included in the fused library, only the resources that will have their definition included in the fused library should have their namespace rewritten to the fused library R class. References to dependencies of the fused library should not have their R class reference changed.

Bug: 426351548
Test: FusedLibraryClassesRewriteTaskTest
Change-Id: I9f2bd66c0a3f33da741cef5a99297a55d63dc923