[Built-in Kotlin] Move built-in Kotlin code to BasePlugin
Previously, built-in Kotlin code was placed in
AndroidPluginBaseServices, which is shared across `BasePlugin`s and
other plugins such as KotlinMultiplatformAndroidPlugin,
FusedLibraryPlugin, and PrivacySandboxSdkPlugin.

This commit moves this code to `BasePlugin.kt` to make it clear that
built-in Kotlin support is available for the core plugins only
(application, library, dynamic-feature, and test), it is not available
for the other plugins.

Bug: 341765853
Test: Existing tests (refactoring only)
Change-Id: Iea9b9786066dff36a1d961aa3b5e26e941f2e29f