[Built-in Kotlin] Delete BuiltInKotlinJvmAndroidCompilation
BuiltInKotlinJvmAndroidCompilation was created as as an interim measure
while we waited for KGP to provide the necessary API to create
KotlinJvmAndroidCompilation required for built-in Kotlin support
(https://youtrack.jetbrains.com/issue/KT-77023 ).

Since the KGP API will be available soon,
BuiltInKotlinJvmAndroidCompilation is no longer needed.

In this commit, we remove BuiltInKotlinJvmAndroidCompilation directly
without waiting for the KGP API to land in KGP because
BuiltInKotlinJvmAndroidCompilation is blocking some users using test
fixtures and screenshot tests.

Bug: 429161295
Test: Updated BuiltInKotlinCompilerPluginTest
Change-Id: Iff95c18f0e18003dc3d056a1cb99d3c71fdfd9b2