Remove excessive checks for lint jars
After fixing findRuleJars to trigger checking for lint jars in more
cases (necessary for IDE Lint), we encountered slow performance on
Windows due to a large number of calls to File.exists() (where the
result is always false).

This change removes checks for lint jars that no longer seem to be
necessary. In particular, I think the code under "library is
LintModelModuleLibrary" has evolved such that it can never trigger
successfully (because LintModelModuleLibrary is an external library, but
the code is checking intermediate build folders). Also, the entire
LintClient.findRuleJars is (usually) skipped in AGP Lint, which provides
its own list of lint jars via --lint-rule-jars.

Bug: 370353954
Bug: 424430560
Test: presubmit
Change-Id: Id0d87d1b64f53d010c5c175f14a143ddb08a9888