[fused lib] Ignore constraints when validating dependencies
Constraints are recognized as dependencies in the `Require transitive dependency inclusion` validation check. This is misleading as for dependencies that have a transitive dependency constraint of itself, and all the parent transitive dependencies are not specified in the include configuration then the check incorrectly fails.

This issue doesn't present itself when there the constraint is only specified in the POM, only in the gradle .module file. We haven't yet built out a test fixture that can generate .module files for dependencies, so the test uses an exisiting Androidx library that produces the issue.

Fixes: 428906893
Test: FusedLibraryClassesVerificationTest
Change-Id: Ib7cd2aefc08379552b691add5d5185fb3a08c128