[fused lib] Avoid triggering assertion when jacoco in R class
When asserting R class fields expected when producing the Fused Library R class the test currently asserts there's just one field 'dependencyLib3_str'. However, when running with coverage enabled, an addition 'jacocoData' field is added to the class (this should be ignored).

This is causing a test failure in studio-coverage (bug linked).

Fixes: 431298236
Test: FusedLibraryClassesRewriteTaskTest assertion more lenient
Change-Id: Ia85d5b86f3745f1d65a24bee527ab1833aff7170