pagealign: add test for unaligned compressed SO
An APK that has an .so file that is stored compressed and is not
zip-aligned at a 16 KB boundary should not report a top-level 16 KB
warning in the APK Analyzer root node.

Before this change, the attached APK will show a warning at the root
but there will be no warnings for any individual SO files (because
the they are valid for 16 KB alignment purposes).

Bug: 432782053
Bug: 382091362
Test: Added an integration test with a specially constructed APK.
Change-Id: Ieb6b8b3b2ce5c01c6086e3edb93a5e774c7aa3bb