adblib: Refactor `AdbDeviceSyncServices`
* Move implementation of most logic related to the `sync`
connection to a new `SyncConnection` class, that is shared
by all operations handlers (`stat`, `recv`, etc.)

* Add a `quit` handler to ensure a `sync` connection is
shutdown in a orderly fashion (i.e. all bytes are sent and
received on both `adblib` and adb server sides). This is
to match the behavior of the native adb client.

* Add ability to use buffered channel for sending and
receiving data for a `sync` connection. This gives callers
more control over buffering of the underlying socket
connection.

* As a side effect of both of the previous changes,
a `sync` service instance needs to be `AutoShutdown.shutdown`
when no longer needed, as opposed to merely
`AutoCloseable.close`. All callers have been updated to use
the new behavior.

Bug: 425918523
Test: Included
Change-Id: I781cf1e2081b4330c83512cd5a134676c01752c2