Enabling clock-type option Wall/Dual for Java/Kotlin Method Tracing
The default method tracing configuration is changed to use only wall
clock for API>=34. The user interface is updated to reflect this change
and to allow users to revert to dual-clock tracing if needed.
For API<34, ART traces are always collected with information from both
the clocks regardless of user selection.

Bug: 428876001
Test: existing tests were updated
Change-Id: I0d1349b0abecc795b517bee1c7363e4f807c5a75