Adds bytecode versioning to Live Edit
Live Edited object instances are now assigned a reference to the
version of class bytecode they belong to.

If changes incompatible with the interface of that object occur in
Live Edit, the object retains its current (compatible) verison of
the class bytecode. If the initial Live Edit bytecode is inocmpatible
with the object instance, no bytecode is assigned and the object
retains pre Live Edit behavior.

This prevents issues where lambda renumbering/similar changes cause
MyClass$lambda$1 to change from a Function0 to a Function3, or from
a SuspendLambda to a Lambda. In both cases, using the latest version of
the bytecode will cause calling code to crash, as the object exposes an
interface [invoke(a, b, c) or invokeSuspend, respectively] which the
current Live Edit bytecode for the class has no implementation for.

In these cases, Live Edit will utilize the last compatible version of
the bytecode to resolve the method calls.

Fixes: 390176327
Test: test/java/c/a/t/d/LiveEdit*Test; LiveEditVersionCheckTest
Change-Id: I02342e70061b246bf3e6dc8dab5ef50a3d466a35