Remove postProcessing DSL for 9.0
Removed PostProcessing interface, PostProcessingBlockOptions,
PostProcessingBlock, ProguardFiles.NO_ACTIONS.

PostProcessingFeatures usage is reduced to only a model for
R8 settings.

CheckProguardFiles was used only when postprocessing block is
there. Now it's not in use.

R8ToolParameters.disableMinification was tight with
postprocesssing.obfuscate. As now there is no such api
need to find way to disable obfuscate in next CL
for JacocoWithMinificationIntegrationTest

Relnote: part of API removal for 9.0
Bug: 265255528
Test: existing
Change-Id: I005a3277c3f18d7f7cb0374dd4b00cb94c606d9c