move VariantBuilder.targetSdk and targetSdkPreview to hidden
Moved the API to hidden so we keep binary compatibility in 9.x
Further optimization like moving the implementation from VariantBuilder
to actual subtypes of GeneratesApkBuilder should be done when the
API is removed.

Relnote: VariantBuilder's targetSdk and targetSdkPreview are now deprecated at the HIDDEN level. Users need to use GeneratesApkBuilder equivalent APIs.
Test: Existing
Bug: N/A
Change-Id: Ifb9ca19461f0732253ef1384a378313054df2223