Add OpenAI Java SDK to maven repo
This CL adds the OpenAI Java SDK to maven/data.bzl. It is
intended for consumption by aiplugin. The source archives
for this jar seem slightly busted (contains an unnecessary
folder "main" at the root) and IntelliJ seems unable to
use it, so they are not added.

The changes to BUILD.maven were auto generated by running
the maven_fetch script.

Bug: 427350534
Test: N/A
Change-Id: I39fabb562d04423cf12a690ef970ab89cd1ea3a2