added consumerProguardFiles to Variant API.
removed equivalent deprecated methods from the VariantScope.
reduce VariantDslInfo API surface to a single methods that is
called from the Variant objects.
Removed checks in the ModelBuilder that could not be performed
any longer due to the late binding nature of the variant API.

Relnote: Added consumerProguardFiles to variant API for Library components.
Test: added unit test.
Bug: N/A
Change-Id: I96da4a6d33b1a4b9f45a29ff80972182d9b3bf72