Flag API tweaks
Based on recent changes to how we use Flags, this commit
makes naming tweaks to the API.

ImmutableFlagOverrides was used for the file-based
boolean flag default value provider but it's not actually
an override, so this is renamed to a simpler FlagValueProvider.

The mutable version is therefore changed to FlagValueContainer.

Overrides implementations are kept as is as they are actually
overrides. The file base provider in onother module will
be renamed approprietely.

This commit also tweaks the naming of the various fields
in Flags to make it more clear what is going on, and adds
more javadoc.

Finally, Flags.getOverrides() did not actually return
all the overrides, so this is renamed to getUserOverrides().

This commit also introduces a new API for Flag override
based on Feature Configuration. This allows unit tests
to simulate various configurations to test different
code paths.

Bug: n/a
Test: existing tests
Change-Id: I1660833ae7c6e1e6333d303d1b197b033a5c531f