Also instantiate type parameters in bounds
When instantiating a function type `∀ X, (B ≼ Box<X>). (B) -> ...` on
argument `Box<A>`, the analysis was only instantiating `B` to `Box<A>`,
and missing `X` to `A`. This change fixes the most common cases, and
defer a proper fix (e.g. taking into account sub-type declarations
such as `IntBox ≼ Box<Int>`) to a future CL.

Bug: 437405527
Test: new
Change-Id: I027168ed9fe45c93ec0edbca44bc9313958f728e