Use `getArgumentForParameter` less
The analysis used to rely on UAST's `getArgumentForParameter` for
matching arguments and parameters in complex cases. But this turns out
expensive sometimes, and `getArgumentForParameter` on calls to local
functions don't always work (discovered in new test
`testLocalFunctionInsideClass`).

This change fixes the analysis to only resort to `getArgumentParameter`
when there are varargs or argument passing by name that might be out
of order.

Bug: 428673464
Bug: 439660443
Test: new
Change-Id: If24d49b90532bfc66923aab4d7987c2f5f540f17