added facility to invoke test suites with additional inputs
Most of the test suite test engine inputs are descrived in the DSL
when adding a test suite.

However, this is not convenient, as some inputs may be a private
contract between the IDE and the test engine. This is particularly
useful for authentication related parameters.

Now IDE can provide additional paramters to the test engine using
a predifined String.Option pointing to a file. The content of the
file is free formed and is a private contract between IDE and
test engine, Gradle or AGP will not attempt to parse it.

Test: Added a integration test.
Bug: N/A
Change-Id: I6208b66e3e7ced1e8643916ea7fc9eba3e85f0d5