Alter the version boundary for AGP versions with built-in Kotlin
Predicting the future is hard.  Predicting the future when other
developers are doing experiments with version numbers is even harder.
Move the AGP_VERSION_WITH_BUILD_IN_KOTLIN constant to somewhere even
further in version number space so that it is not before
9.0.0-dev (with our somewhat idiosyncratic convention that our -dev
version, never released to users, is *newer* than any other preview
version).

This is not a dangerous change because as soon as we *do* have a
version of AGP with built-in Kotlin, template tests (at least) will
start failing.

Bug: none filed
Test: existing
Change-Id: Ia533a2b45bc00844191111ab89025b43430458be