Exclude tools:overrideLibrary from <uses-sdk> check.
Instead of throwing error when <uses-sdk> is found in the
manifest, we will check if it contains sdk controlling
attributes. This will advise users to remove the <uses-sdk>
tag if such attributes are found. If user is using the
tools:overrideLibrary, it will advise removing forbidden
attributes while allowing the tools one.

Bug: 440557965
Test: ManifestMerger2SmallTest
Change-Id: I1da30d7e3a2737c50c3e828e6f64cca1e1c84f62