Ensure that JAVA_RES added through artifacts API are used.
There are two ways to add Java resources to a gradle project
through the Variant API. The first one, is using the Sources API
and doing something like sources.resources?.addGeneratedSourceDirectory.
The second is the use the scoped artifact JAVA_RES and append to
it. Ensure that both ways work and that the generated java resources
are invoked when merging.

Note that java resources added through the Sources variant APIs are
not visible to the JAVA_RES artifact except in their merged state.
Similarly the java resources added as an artifact are not visible to
the Sources.resources.all() API.

Test: Added regression test.
Fixes: 437906190
Relnote: N/A
Change-Id: Icb1e2b6034e3c22e086aaaaec4b2b462648f444b