Use KGP API to create KotlinCompilation instances
For built-in Kotlin to work correctly, AGP needs to create
KotlinCompilation instances.

Previously, since there wasn't an API to do this, we had to rely on a
private KGP API that requires passing the old BaseVariant object.

Now that we have updated AGP to depend on KGP 2.2.10 where the proper
API is available (KT-77023), this commit will use that API to create
KotlinCompilation instances.

This means that screenshot tests, test fixtures, compiler plugins are
now fully supported with built-in Kotlin. The `Kotlin not configured`
issue that was appearing in the IDE for screen tests, test fixtures, or
for the main code when `android.newDsl=true` is now resolved.

Bug: 429981132
Test: Updated BuiltInKotlinCompilerPluginTest
Change-Id: I56bee3739b613dce6b33a11e6127b0a58b006b5e