disallow adding a Provider<*> to AndroidSourceSet
Setting a Provider<*> on srcDir or srcDirs AndroidSourceSet APIs
does not carry task dependency due to the old Variant API that is
resolving these directories too early. Even when that will be resolved
it is not possible for the IDE to know if this is a generated
or a static directory so it's better to guide users to use the
new Variant API that have dedicated APIs.

Test: Added integration test.
Fixes: 333831734
Relnote: Disallow usages of Provider<*> with AndroidSourceSet.srcDir APIs
Change-Id: Iaef96823d6393f85d9a571a56caed22adc5883ec