Update Journeys template to configure a test suite
A 'test suite' field has been added to the Journeys template, to allow
users to specify which suite the Journey file should be added to.

If the chosen test suite doesn't exist, the template will automatically
configure one, applying the configuration needed to run Journeys, i.e.
adding the dependency on the test engine, setting the input parameters,
targeting the current variant, etc.

Test: snapshot tests in ag/35827038
Bug: 433523178
Change-Id: I99c258c8ea19f738292e8f9110d5812373b1793e