introduce 'includeAndroidResources' in host test and test suites.
Added the flag in HostTestBuilder as settable, read-only value in
HostTest. Added the flag in TestSuite's host jar definition (variant
API to follow). Removed access to the global extension flag so
internal users are not tempted to use that value.

Fixes: 382601039
Test: Added integration test
Relnote: added Variant API for included android resources in host test and test suites.
Change-Id: I09551ba00267b6c27ddcf28513190de1f985f128