Allow heap size override for screenshot tests
This change introduces a standard Gradle property,
`android.compose.screenshot.maxHeapSize`, to control the memory for the
screenshot tasks. If this property is set by a user (e.g., in their
`gradle.properties` file), its value will be applied as the
`maxHeapSize` for both the validation and update screenshot tasks.

Bug: 44404802
Test: Manually verified on the project from the bug report.
Change-Id: I6acabafcde04afabfe1fea46b59bc631c1f464ca