Add adapter and consumer layer to support implementation of a backend agnostic engine.
This change introduces a new adapter-consumer mechanism where each backend is linked to a unique adapter. The task of this adapter is to accept any results coming from the backend and generate JourneyRunEvent according to the received result. This event is then consumed by aconsumer to produce encoded proto messages in stdout. This encoded message will be read on Studio side to generate journey results.

Test: JourneysConnectedTest
Bug: 433521353
Change-Id: Ib772a5335c6b9476e114dc419522a1fafc092f2b