add internal APIs to handle KSP/KAPT dependencies
KAPT and KSP are both generating java code and KAPT needs to depend
on KSP to generate java stubs for KSP generated interfaces. On top
of that, bunch of other generators exist in the AGP or in the wild
that can easily create circular dependencies when injecting 'all'
sources.

Added the ability to store KAPT and KSP generators separately and
an API to query all generators without those. This new internal
API can be used by KSP to avoid circular dependencies.

Added a test to simulate these relationship and have an
easier way to ensure proper API usages and stability

Test: integration test
Bug: N/A
Change-Id: I3e3efb7b7f4568b6d679bb8943652be4cceef1d0