Remove usage of internal Problems API
InternalProblems service was used because it provided better support for
additional data in the events. In public api it either was very confusing
or didn't work at all.
As now usage of this internal service causes problems, change to recommended
way of providing additional data via public problems API.

Bug: N/A
Test: AndroidProblemsReporterTest, all integration tests dealing
with sync issues.

Change-Id: Ie7667c293022114761dc17abb8d96471cb6c39d7