Write the hw.screen property if screen-type is notouch
There is not a 1:1 mapping between the values hw.screen expects, and the
values that <d:screen><d:screen-type> gives. Currently, the device.xml
for AI glasses sets screentype to notouch with the reasonable
expectation that this will translate to hw.screen=no-touch.
This makes that the case, while not handling the other values which are
not of any importance to the emulator currently.

Bug: 478340502
Test: Create new AI Glasses avd, confirm hw.screen=no-touch is set
Change-Id: Ib059e1334aa3ae1b0ee7670759e6ad8bb3c0d354