Fix injected manifest used in lint tests
The injected manifest is used in lint tests to to inject the minSdk and
targetSdk from the Gradle build file into the merged manifest. The
manifest merger started rejecting the injected manifest, due to various
issues.

We fix the LintCliClient code to use major version numbers, and provide
the injected manifest as an overlay manifest. We also throw an exception
in unit tests so we can catch manifest merging problems like this. We
skip handling exceptions (handleDetectorError) in unit tests.

Added `TestLintTask.allowManifestMergerErrors(Boolean)` so that tests
can ignore manifest merging problems.

Bug: 448851433
Test: fixed existing
Change-Id: I44d22f0b337bf02c7998136519b0c67c9cb84444