Fix a crash caused by `AdbActivityManagerException`
Addresses a crash report caused by an unhandled
`AdbActivityManagerException`.

The method `ConnectedDevice.useAppInfoForProcessProperties` calls
`AdbActivityManagerServices.capabilities`, which can throw this
exception. While the exact cause is unknown, it does not appear to be
related to the 'activity' service being unavailable or the
'capabilities' command being unsupported, as those states are already
handled by `ActivityManager.retryUntilDeviceReady`.

Bug: 437306401
Test: existing
Change-Id: I4592d1c3830eb7f02d524ab2686309fcd08d3be4